requirements.txt
alembic
blinker
chardet
click
Flask
Flask-Dance
Flask-DebugToolbar
Flask-Login
Flask-Migrate
Flask-OAuth
Flask-OAuthlib
Flask-SQLAlchemy
Flask-WTF
httplib2
idna
itsdangerous
Jinja2
lazy
Mako
MarkupSafe
oauth2
oauthlib
python-dateutil
python-editor
requests
requests-oauthlib
six
SQLAlchemy
SQLAlchemy-Utils
urllib3
URLObject
Werkzeug
wincertstore
WTForms
pillow
email_validator